requirements.txt
blinker==1.9.0
boto3==1.35.81
botocore==1.35.81
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
urllib3==2.2.3
Werkzeug==3.1.3
